Fix some Windows warnings

+++ storage/innobase/handler/ha_innodb.cc
@@ -11823,7 +11823,7 @@ err_col:
 			err = row_create_table_for_mysql(
 				table, m_trx, false,
 				(fil_encryption_t)options->encryption,
-				options->encryption_key_id);
+				(uint32_t)options->encryption_key_id);
 		}
 
 		DBUG_EXECUTE_IF("ib_crash_during_create_for_encryption",
